requirements.txt
streamlit>=1.32.0,<1.45.0
streamlit-drawable-canvas==0.9.3
sqlalchemy>=2.0.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.0.0
